Fix usage of Requirement class.
Requirement class automatically generate in its constructor the formula name using the name of deriving class.
In the case of jubatus, if the requirement name is ZooKeeperRequirement, then it sets @name = 'zookeeper'.
We can also set it manually in constructor, but naming with prefix 'Requirement' or 'Dependency' is maybe the homebrew way (I think so because such automatic behavior exists), so obeyed it.